PCSX2/EEcore:  * Now using SSE for all hardware register reads and writes (mainly MFIFO stuff) [don't expect a speedup, really -- its more of a code simplification in this case].  * [refactoring] Changed the EE Memory (vtlb) to use the u128 type instead of u64 for the 128-bit loads/stores (see mem128_t typedef)
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@3626 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ pcsx2/MemoryTypes.h
@@ -35,4 +35,4 @@ typedef u8 mem8_t;
 typedef u16 mem16_t;
 typedef u32 mem32_t;
 typedef u64 mem64_t;
-typedef u64 mem128_t;
+typedef u128 mem128_t;
